dashing implementation update
The most annoying ability known to man.

I have set dashing to work on the "enemy path test map". I realized that we would need to actively set the map we want the abilities to occur in. The only good thing from spending hours on this

Repressed all use of shift now. z to dash

Dash system summary
Unless someone can come up with + implement a better system for this &&%%#$ ability here is what I could implement

-Player can dash every 3 seconds

We can discuss this in a gen/prog meeting if you want.