300MB is too much
Mass deletion spree

>300MB -> 90MB

There is the occasional file error even though we are not directly using it.
We may need some testing for this and will have to deal with it on a case by case.
So do play the game again to see if any file not found errors pop.