L2 Ability and Stunning Mech Fixes
I think it would be fitting to have upgrades for the more robotic cat in a robotics lab. So I just went crazy with it

-increased base bullet speed

-fixed stunned NPC from default

-Left chamber gives bigger bullet:
 G to shoot

-Right gives ranged bullet: J to shoot

-Center gives more ammo

-top right increases stun duration

-Player can still use regular bullet: H tp shoot. though maybe we can take it away.

thorough testing required to ensure all these work as intended

I added comments for Max and Alan on the Chamber events. Will also message to hash out any needs.