Just L2
a drag but its all good now. Same comments with slight change but mostly same as before but [copypasted]

-LOS for L2 should be working like so;

if player does not shoot the spotter before getting spotted or before continuing after text-> Player gets caught is spotted, loses a 'life' and sent back to start of map.

if player shoots spotter before getting spotted but is still spotted or before continuing text-> text shows but spotter is stunned. Therefore, player does not lose a life and is not sent back to start. They can complete the text and carry on[if they move quickly].

Requires Testing